Fixes #33. Changes the way the installed_paths are set.
Changes the behavior of registering installed_paths in PHP_CodeSniffer 3.x. It will now work in the same way as was required for PHP_CodeSniffer 2.x. It will record the directory ABOVE the directory containing the coding standard. The only exception is when the root project itself provides the standard in the root of its project.