:books: README: replace maintenance badge
The `year` part of the Shields.io `maintained` badge needs to be updated with the current year for it not to show as "No" (unmaintained/abandoned).

I'd like to suggest using a different badge - "Last commit" - instead which doesn't need manual updating each year and still gives a reasonable indication of the maintenance status of the project.

Refs:
* https://shields.io/#/examples/other