Use the PHP version used by Composer
Possible not the best way to do this, but it does work, so consider this a proof of concept, if not the solution.

The `getPhpExecCommand()` method in the `Composer\EventDispatcher\EventDispatcher` class basically does what is needed: find the PHP executable used by Composer and turn it into an executable command.

Unfortunately, that method is `protected`, making it difficult to call that method from within the plugin.

This now copies that method into the plugin and uses it to create the command to pass on to the `ProcessExecutor`.

The copied method has some minor modifications to comply with the coding standards used within this project.

Based on the proof of conflict referenced in 79, I can confirm that this would solve the issue.