Readme: add section about including this plugin from an external PHPCS standard
When an external PHPCS standard itself is dependent on one or more other external standards, it can hep to have this plugin as a `require`ment for the external standard itself.

This commit adds a section about this feature, when to use it, why to use it and how to use it.

Includes minor grammar fix in existing text.