Verify the `installed_paths` after save
As most problems reported are to do with paths not being set correctly, let's verify success within the plugin, independently of PHPCS.

To that end, a new function `verifySaveSuccess()` has been added.
This function takes the paths which were to be saved and compares then to the paths which are actually set in PHPCS after the save to determine whether or not the plugin was successful and changes the exit code if this is not the case.

If there was a mismatch in paths and `verbose` mode is turned on, an information message will be displayed to help debugging.

At this moment I don't have a test case for this new functionality, however, I imagine it will come in useful with future bug reports.