Tests: create zip package artifact for the current plugin code
In normal circumstances, a `composer install` will download a version of the plugin from Packagist.
This will not work for the tests however, as the development version of the plugin being tested will generally not be available on Packagist.

With this in mind, we need a work-around to ensure that the tests are run with the **current** version of the plugin as per the last commit in the current branch.

I've implemented this by using the `repositories - artifact` feature of Composer.

This works as follow:
* In the test bootstrap, ahead of the test run, the plugin is zipped up as a package and placed in the `tests/artifact` directory.
* Whenever a `composer install/update` command is run within the test suite, the `repositories-artifact` key is injected into the `composer.json` when the `composer.json` file is written to the temp directory used for the test.
* With this in place, Composer will install the plugin from the zipped file in the `tests/artifact` directory instead of downloading from Packagist.

Notes:
* The version number to be used for the zip artifact of the plugin is defined in the test bootstrap file.
    This number may need updating once in a while to ensure it is higher than the latest released version of the plugin.
* In the `CreateComposerZipArtifacts` class, three properties are used to determine which files to exclude from the package - `$excludedFiles`, `$excludedExtensions` and `$excludedDirs`.
    Depending on new files/directories being introduced in the package, these lists may need to be updated once in a while.

Ref: https://getcomposer.org/doc/05-repositories.md#artifact