TestCase: add helper method for on-the-fly composer.json creation
This adds a test helper method, which, given an array with a Composer configuration and a target location, will on-the-fly create a `composer.json` file for use in a test.

The method will automatically add two additional settings to the `composer.json` file:
* It will add a `install-codestandards` script to run the plugin on demand.
* It will add the `allow-plugins` permission, which is needed for Composer 2.2.

By adding these keys automatically, this kind of "noise" is not needed in the actual test classes.

The method is `static` so as to allow for it to be used in `setUpBeforeClass()`/`tearDownAfterClass()` fixtures.