Tests: add initial baseline test
These tests verify:
- That the plugin can be installed and functions correctly with the full range of supported PHPCS versions.
    While the test is not run against the full range of supported PHPCS version each time, due to the random PHPCS version selection, all supported PHPCS versions will be tested over time in CI.
- That the plugin runs when installed.
- That the PHPCS native standards are the only recognized standards when no external standards are available.
- That no `CodeSniffer.conf``file gets created when no external standards are found.

These tests will be run in both a Composer global as well as a Composer project local environment.

The tests use a dataprovider which will randomly select two PHPCS versions compatible with the PHP version on which the test is being run + PHPCS `dev-master` + PHPCS `4.x` (dev version for the next major).

As which PHPCS versions are compatible with which PHP versions needs quite some logic, a separate helper class `PHPCSVersions` is introduced, which encapsulates that knowledge.

The `PHPCSVersions` class contains a number of (`static`) methods as helpers. These can be used in any test class:
* `get()`: retrieves an array with a specific number of PHPCS versions valid for the current PHP version.
* `getHighLow()`: retrieves an array of the highest and lowest PHPCS versions valid for the current PHP version.
* `getHighLowEachMajor()`: retrieves an array of the highest and lowest supported PHPCS versions for each PHPCS major (and valid for the current PHP version).
* `getRandom()`: gets a random PHPCS version which is valid for the current PHP version.
* `toDataprovider()`: converts a version array to an array suitable for use as a PHPUnit dataprovider.
* `getSupportedVersions()`: retrieves an array with all PHPCS versions valid for the current PHP version.
* `isNextMajorSupported()`: determines whether the current PHP version is supported on the "next major" branch of PHPCS.
* `getStandards()`: retrieves an array of the names of the PHPCS native standards for a particular PHPCS version.

Generally speaking, the `dev-master` and `4.x` branch will not be included in the retrieved ranges, but can be specifically requested by passing optional parameters.

This `PHPCSVersions` class will need semi-regular updates when new versions of PHPCS are released and new PHP versions are released.