Tests: set up a fake PHPCS standard as a fixture and create an artifact of it
The tests in the `RegisterExternalStandardsTest` class contain quite some work-arounds to handle changes made over time to the external PHPCompatibility standard.

The `PHPCompatibility` standard was chosen for use in the tests as it is one of the rare standards which still supports both PHPCS 2.x as well as 3.x.

All the same, to improve the stability of the tests and to simplify the actual test code, using "fake" PHPCS standards set up as fixtures as part of this test suite will make life easier.

This commit creates the initial setup for that, by:
* Adding code to the `CreateComposerZipArtifacts` class to handle creating Composer package zip artifacts for each subdirectory of the `tests/fixtures` directory.
* Calling that code from the test bootstrap file to ensure all test artifacts are zipped up and ready for use ahead of running the tests.
* Adding an initial "fake" `DummySubDir` standard as a test fixture.
* Adding a `README` file to the `tests/fixtures` directory with the rationale and a list of available fixtures.