Plugin: improve feedback
The `Nothing to install or update` message is confusing, as for an end-user, it may not be clear that this message relates to the PHPCS plugin.
The message as-is could be interpreted to relate to Composer itself finding _nothing to install or update_ instead.

I'd like to suggest changing that message to be more directly related to this plugin for clarity.