Tests: add new `PlayNiceWithScriptsTest`
This new test class tests that the plugin does not block other post install/update scripts from running.

I have confirmed that this test as-is would fail in combination with the `0.6.0` and `0.6.1` versions of the plugin and would pass with `0.5.0`, `0.6.2` and current.
For confirmation see this build against `0.6.0`: https://github.com/Dealerdirect/phpcodesniffer-composer-installer/actions/runs/1787983161 (only run against Composer 1 as `0.6.0` was not compatible with Composer 2 yet).

Notes:
* This test is about Composer and the plugin, so does not need to be tested against multiple PHPCS versions.
* The behaviour also shouldn't differ between a global vs local Composer install, so only testing one type.

This test class covers the following bug previously reported:
* Dealerdirect/phpcodesniffer-composer-installer PR 105