Tests: add new `InstallUpdateEventsTest`
Test that the plugin is hooked into the right events and doesn't get triggered when those events are blocked (`--no-plugins`).

These tests verify that:
- The plugin gets triggered on install/update/require/remove/reinstall events.
- The plugin doesn't run when Composer is run with `--no-plugins`.
- The plugin does run when Composer is run with `--no-scripts` (since Composer 2.1.2).
- Can be run on-demand via a custom script.

This test is about Composer and the plugin, so does not need to be tested against multiple PHPCS versions.
The behaviour also shouldn't differ between a global vs local Composer install, so only testing one type.

Note: the `composer reinstall` test will only run on Composer 2.2.6+ as prior to that release plugins would not run on that command.

This test class covers the following bug previously reported:
* Dealerdirect/phpcodesniffer-composer-installer issue 4
* Dealerdirect/phpcodesniffer-composer-installer PR 5