TestCase: new `recursiveDirectoryCopy()` method
New helper method to recursively copy a (fixture) directory with all its files to another location, typically, one of the system temp test directories.

If the `$srcDir` contains a `composer.json` file, it will get the same `TestCase::writeComposerJsonFile()` special treatment, meaning that the artifacts directory will be injected, as well as plugin permissions and more.