Tests: add new `RemovePluginTest`
Test that the plugin does not throw errors when the plugin is still in memory, but PHPCS and the plugin are uninstalled (Composer 1.x).

These tests verify:
- That the plugin exits with error code 0 when PHPCS and the plugin are uninstalled due to a no-dev install.
- That the plugin exits with error code 0 when the standard which required the plugin is uninstalled.

This test is about Composer and the plugin, so does not need to be tested against multiple PHPCS versions.
The behaviour also shouldn't differ between a global vs local Composer install, so only testing one type.

Note: the behaviour of Composer changed for this in Composer 2.0.0, but still not a bad thing to safeguard this for Composer 2.x as well, in case of regressions in Composer.

This test class covers the following bug previously reported:
* Dealerdirect/phpcodesniffer-composer-installer issue 103
* Dealerdirect/phpcodesniffer-composer-installer PR 104