GH Actions: fix builds on Windows with PHP 8.2
Looks like the `zip` extension is not available by default on [Windows with PHP 8.2](https://github.com/shivammathur/setup-php/wiki/Php-extensions-loaded-on-windows-2019#php-82), which is causing those builds to fail as the zip extension is needed in the test bootstrap.

Looking at the `composer.json` file, we previously identified the `json` and `zip` extensions as `dev` requirements, so let's ensure they will always be [installed and enabled in CI](https://github.com/shivammathur/setup-php#heavy_plus_sign-php-extension-support).