GH Actions: update PHP versions in workflows
PHP 8.2 has been released today :tada: and the `setup-php` action has announced support for PHP 8.3, so adding PHP 8.3 to the matrix and no longer allowing PHP 8.2 to fail the build.

Builds against PHP 8.3 are still allowed to fail for now.

Includes updating the `PHPCSVersions` class for the test suite to contain a PHP `8.3` block and updating the PHPCS table to include the latest releases. (we should automate that)