README: updates related to the 1.0 release
The readme still contained various references to "minors being treated as majors below v1.0".

This information was relevant prior to the 1.0 release (and slightly after), but the `1.0` version has been out for over a year now and adoption is good with approximately 70% of users having updated to 1.0 so far.

This commit updates the README text to simplify the information related to this.

Refs:
* https://packagist.org/packages/dealerdirect/phpcodesniffer-composer-installer/stats