Tests: allow for tests to use root package
The zip package we create in the test bootstrap to allow for Composer to use the root version of this package when installing needs to have a higher version number than any released version. The version number, however, should also be compatible with the current major.

This updates the version number in the test bootstrap file.