Tests: run against the *new* PHPCS 4 branch
The PHP_CodeSniffer 4.0 branch has been completely rebuild. See https://github.com/PHPCSStandards/PHP_CodeSniffer/issues/120 for more detailed information about the "why".

In practice, this means that to test against the upcoming PHPCS 4.0 release, we need to test against the `4.x` branch now, not the old `4.0` branch.

A 4.0 beta/RC release is expected in the next few days. Once that is out, a test period starts and depending on whether problems are found or not, the 4.0.0 release will be tagged about a month later (or later if significant problems were found).

When 4.0.0 gets tagged, the PHPCS `master` branch will be renamed to `3.x` and the `4.x` branch will be the new "main" branch for PHPCS, so more changes will be needed, but that's for later.

Includes updating to the `DemoSniff` test fixture setup as PHPCS 4.0 will start enforcing for sniffs to implement the `Sniff` interface.