Add CHANGELOG.md file
The contents is based on the existing releases, with only minor ordering tweaks to follow the "Keep a changelog" categorization.

This commit includes removing the configuration and script for the previously used changelog generation tooling.

Potherca and me discussed this today and decided to go for manually crafted changelogs from now on as:
1. The previous changelogs contained a lot of "noise" about non-user facing changes.
2. The changelog generation tooling did not run on Windows, making it less useful as not all maintainers could run it.