Fix "The provided cwd "../squizlabs/php_codesniffer" does not exist."
I'm very curious why the reported issue only appears to happen with a Drupal setup (and I cannot reproduce it, nor do our tests fail). All the same, I'm hoping this will fix it.

Note: yes, `realpath()` can return `false`, but that shouldn't be possible in this case as we only call it after we've already checked `isPHPCodeSnifferInstalled()`.