Modernize: add constant visibility
For the tests, all visibility modifiers have been set to `private`.
For the `src`, all visibility modifiers have been set to `public` as anything else could be considered a BC-break.

I suggest we open a ticket to change the constant visibility for the class constants in the `Plugin` class to `private` in the next major.